Create Collatz_1-7.py
Evirir the dragon has a list of n integers a1, a2,…, an. Evirir will repeat the following procedure k times: 

For each integer ai in the list:
If ai is even, replace ai with (ai)/2. 
Otherwise, replace ai with 3ai+1.

In the end, what is the sum of all integers in the list?

Input Format:
The first line consists of two integers n
 and k (1≤n≤1000, 1≤k≤1000). Then n
 integers follow: a1, a2…, an (1≤ai≤104).

Output Format:
Output a single integer, the sum of the integers in the final list.