Create Tichu_4-6.py
You have N cards in your hand. N−K cards have integers on them. The remaining K cards are wild cards that can each represent a card of any integer you choose. A run of length m is a sequence of m cards with the integers (a, a+1, a+2, …, a+m−1) for some integer a. For example, the sequences (7, 8, 9), (1, 2, 3, 4, 5), and (6) are runs, but (2, 3, 5, 6), (1, 6), and (4, 3, 2, 1) are not runs.

What is the longest run you can make with these cards? You may rearrange the cards in your hand however you like to form this run.

Input Format
The first line consists of an integer N (1≤N≤105), followed by another integer K (0≤K≤N). The second line contains N−K integers, representing the cards in your hand that aren’t wild cards: C1,C2,…,CN−K (1≤Ci≤10^9).

Output Format
Output a single integer, the length of the longest run you can form from the cards you have.